requirements.txt
python-dotenv
streamlit
crewai
crewai_tools
langchain-mistralai
langchain_openai
